+++ lib/iord/controller.rb
@@ -1,4 +1,6 @@
 require 'active_support/concern'
+require 'hooks'
+require 'iord/defaults'
 require 'iord/resource_info'
 require 'iord/resource_url'
 require 'iord/crud'
@@ -10,6 +12,7 @@ require 'iord/resolver'
 module Iord
   module Controller
     extend ActiveSupport::Concern
+    include Iord::Defaults
     include Iord::ResourceInfo
     include Iord::ResourceUrl
     include Iord::Attriutes
@@ -17,8 +20,10 @@ module Iord
     include Iord::Fields
 
     included do
-      append_view_path Iord::Resolver.new
+      include Hooks
 
+      append_view_path Iord::Resolver.new
+      define_hook :before_set_resource
       before_action :set_resource
 
       cattr_accessor :resource_based_actions, instance_accesssor: false do
@@ -41,6 +46,7 @@ module Iord
 
     protected
     def set_resource
+      self.run_hook :before_set_resource
       return unless self.class.resource_based_actions.include? params[:action].to_sym
       if params[:id]
         @resource = resource_class.find params[:id]