+++ test/dummy/app/controllers/subcomments_controller.rb
@@ -2,10 +2,12 @@ require 'iord/controller'
 require 'iord/nested'
 require 'iord/json'
 
-class CommentsController < ApplicationController
+class SubcommentsController < ApplicationController
   include Iord::Controller
   include Iord::Nested
   include Iord::Json
+
+  defaults resource_class: Comment
 
   def index_attrs
     %i(author_name)