+++ test/factories/comment.rb
@@ -7,5 +7,11 @@ FactoryGirl.define do
     content { Faker::Lorem.paragraphs(1) }
     commentable { build(:product) }
   end
+
+  factory :subcomment, class: 'Comment' do
+    author_name { Faker::Name.name }
+    content { Faker::Lorem.paragraphs(1) }
+    commentable { build(:comment) }
+  end
 end