+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* `Iord::Json` now works with `Iord::Nested`
+* Added `Iord::Defaults`
+
 ## 0.3.0 - Nested resources
 * Added module `Iord::Nested` to handle nested resources